// src/domain/models/Examenes.Model.ts
export interface ExamenesModel {
    id:                 number;
    consults:           string;
    exam_date:          Date;
    mucosa:             string;
    piel:               string;
    conjuntival:        string;
    oral:               string;
    sis_reproductor:    string;
    rectal:             string;
    ojos:               string;
    nodulos_linfaticos: string;
    locomocion:         string;
    sis_cardiovascular: string;
    sis_respiratorio:   string;
    sis_digestivo:      string;
    sis_urinario:       string;
    status:             number;
}